Moved HatchControls to be on the same level as all the other teleop commands.
It was in HatchControls, but ClimberControls, DriveControls, and CargoControls were on a different level in the file system. I fixed it, and fixed imports elsewhere, so everything still works.
